+++ lib/k8s.w
@@ -1,0 +1,10 @@
+class Pod {
+  init(podfile: str) {  }
+}
+class Service {
+  init(pod: Pod){}
+}
+class Ingress {
+  addRule(path: str, service: Service){}
+  onStart(fn: inflight (str):void) { }
+}
