+++ lib/vite.w
@@ -1,0 +1,4 @@
+class Website {
+  init(configuration: str){}
+  onStart(fn: inflight (str):void) { }
+}
